+++ src/main/java/dnd/manager/app/model/BonusTraitObjectId.java
@@ -3,6 +3,6 @@ package dnd.manager.app.model;
 import java.io.Serializable;
 
 public class BonusTraitObjectId implements Serializable {
-    private Integer traitId;
-    private Integer objectId;
+    private Long traitId;
+    private Long objectId;
 }
