+++ src/main/java/dnd/manager/app/model/BonusTraitStatId.java
@@ -2,6 +2,6 @@ package dnd.manager.app.model;
 
 import java.io.Serializable;
 public class BonusTraitStatId implements Serializable {
-    private Integer traitId;
-    private Integer statId;
+    private Long traitId;
+    private Long statId;
 }
